requirements.txt
numpy
matplotlib.pyplot
scipy
pandas
streamlit